requirements.txt
playwright==1.44.0
pytest==8.2.2
pytest-playwright==0.5.0
pytest-html==4.1.1
pytest-metadata==3.1.1
Faker==25.2.0
python-dotenv==1.0.1
